requirements.txt
alabaster==1.0.0
altair==5.5.0
attrs==25.3.0
babel==2.17.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
Cython==3.1.2
docopt==0.6.2
docutils==0.21.2
filelock==3.18.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
hf-xet==1.1.3
huggingface-hub==0.32.5
idna==3.10
imagesize==1.4.1
indic-nlp-library-itt==0.1.0
IndicTransToolkit @ git+https://github.com/VarunGumma/IndicTransToolkit.git@37483c962c486f2f61d8b3f659c0a1064c55ac6c
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
langdetect==1.0.9
lxml==5.4.0
MarkupSafe==3.0.2
Morfessor==2.0.6
mosestokenizer==1.2.1
mpmath==1.3.0
narwhals==1.42.0
networkx==3.5
numpy==2.3.0
openfile==0.0.7
packaging==24.2
pandas==2.3.0
pillow==11.2.1
portalocker==3.1.1
protobuf==6.31.1
pyarrow==20.0.0
pydeck==0.9.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
roman-numerals-py==3.1.0
rpds-py==0.25.1
sacrebleu==2.5.1
sacremoses==0.1.1
safetensors==0.5.3
sentencepiece==0.2.0
six==1.17.0
smmap==5.0.2
snowballstemmer==3.0.1
Sphinx==8.2.3
sphinx-argparse==0.5.2
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
streamlit==1.45.1
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tokenizers==0.21.1
toml==0.10.2
toolwrapper==2.1.0
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tornado==6.5.1
tqdm==4.67.1
transformers==4.52.4
typing_extensions==4.14.0
tzdata==2025.2
uctools==1.3.0
urllib3==2.4.0
